requirements.txt
requests
python-telegram-bot
python-dotenv
Pillow
filelock